requirements-pip-freeze.txt
aiohttp==3.8.3
aiosignal==1.3.1
ape-alchemy==0.5.4
ape-hardhat==0.5.4
ape-polygon==0.5.1a1
ape-vyper==0.5.2
asttokens==2.2.1
async-timeout==4.0.2
attrs==22.2.0
backcall==0.2.0
base58==1.0.3
bitarray==2.6.2
cached-property==1.5.2
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
commonmark==0.9.1
cytoolz==0.12.1
dataclassy==0.11.1
decorator==5.1.1
Deprecated==1.2.13
eip712==0.1.7
eth-abi==3.0.1
eth-account==0.7.0
eth-ape @ git+https://github.com/sajal/ape.git@8a2beb3ebd39fdb42317f711cda6394177d02968
eth-bloom==1.0.4
eth-hash==0.3.3
eth-keyfile==0.6.0
eth-keys==0.4.0
eth-rlp==0.3.0
eth-tester==0.7.0b1
eth-typing==3.2.0
eth-utils==2.1.0
ethpm-types==0.3.16
evm-trace==0.1.0a15
exceptiongroup==1.1.0
executing==1.2.0
frozenlist==1.3.3
greenlet==2.0.2
hexbytes==0.3.0
idna==3.4
ijson==3.2.0.post0
importlib-metadata==6.0.0
iniconfig==2.0.0
ipfshttpclient==0.8.0a2
ipython==8.9.0
jedi==0.18.2
jsonschema==4.17.3
lru-dict==1.1.8
matplotlib-inline==0.1.6
morphys==1.0
msgspec==0.12.0
multiaddr==0.0.9
multidict==6.0.4
mypy-extensions==0.4.3
netaddr==0.8.0
numpy==1.24.1
packaging==20.9
pandas==1.5.3
parsimonious==0.8.1
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
pluggy==1.0.0
prompt-toolkit==3.0.36
protobuf==4.21.12
ptyprocess==0.7.0
pure-eval==0.2.2
py-cid==0.3.0
py-ecc==6.0.0
py-evm==0.6.0a1
py-geth==3.10.0
py-multibase==1.0.3
py-multicodec==0.2.1
py-multihash==0.2.3
pycparser==2.21
pycryptodome==3.17
pydantic==1.10.4
pyethash==0.1.27
pygit2==1.10.1
PyGithub==1.56
Pygments==2.14.0
PyJWT==2.6.0
PyNaCl==1.5.0
pyparsing==3.0.9
pyrsistent==0.19.3
pysha3==1.0.2
pytest==7.2.1
python-baseconv==1.2.2
python-dateutil==2.8.2
pytz==2022.7.1
PyYAML==6.0
requests==2.28.2
rich==12.6.0
rlp==3.0.0
semantic-version==2.10.0
six==1.16.0
sortedcontainers==2.4.0
SQLAlchemy==2.0.0
stack-data==0.6.2
tomli==2.0.1
toolz==0.12.0
tqdm==4.64.1
traitlets==5.8.1
trie==2.1.0
typing_extensions==4.4.0
urllib3==1.26.14
varint==1.0.2
vvm==0.1.0
watchdog==2.2.1
wcwidth==0.2.6
web3==6.0.0b7
websockets==10.4
wrapt==1.14.1
yarl==1.8.2
zipp==3.12.0
